requirements.txt
django
djangorestframework
djangorestframework-gis
